+++ src/constants.js
@@ -9,11 +9,7 @@
 
 import aurbrowser from '../package';
 
-const AUR_BASE_URL = 'https://aur.archlinux.org';
-
-export default {
-  VERSION: aurbrowser.version,
-  AUR_BASE_URL: AUR_BASE_URL,
-  AUR_RPC_URL: `${AUR_BASE_URL}/rpc/?v=5`,
-  CORS_PROXY: 'https://cors.aurbrowser.tk/'
-};
+export const VERSION = aurbrowser.version;
+export const AUR_BASE_URL = 'https://aur.archlinux.org';
+export const AUR_RPC_URL = `${AUR_BASE_URL}/rpc/?v=5`;
+export const CORS_PROXY = 'https://cors.aurbrowser.tk/';