+++ src/views/package.vue
@@ -177,7 +177,7 @@
 </template>
 
 <script>
-  import CONSTANTS from '../constants';
+  import * as CONSTANTS from '../constants';
   import aur from '../aur';
   import utils from '../utils';
   import Prism from 'prismjs';